Create  Arduino code that will turn an LED on when a button is pressed
This code sets up two pins: one for a button and one for an LED. It then uses the digitalRead() function to read the state of the button. If the button is pressed, the code uses the digitalWrite() function to turn on the LED. Otherwise, it turns off the LED. You can modify the pin numbers and the behavior of the LED as needed for your project.